Move preseeding code to a separate package under image/preseed.
